chore: Move keyring interfaces to specific file
